+++ src/main/java/core/dao/impl/UserDaoImpl.java
@@ -1,5 +1,6 @@
-package core.dao;
+package core.dao.impl;
 
+import core.dao.UserDao;
 import core.db.Storage;
 import core.lib.Dao;
 import core.model.User;
