+++ src/main/java/core/service/impl/UserServiceImpl.java
@@ -1,9 +1,11 @@
-package core.service;
+package core.service.impl;
 
 import core.dao.UserDao;
 import core.lib.Inject;
 import core.lib.Service;
 import core.model.User;
+import core.service.UserService;
+
 import java.util.List;
 
 @Service
